refactor ServiceCollectionExtensions
StorageAccountExtensions don't need to exist in their own file.

Related to: #319
